requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
numpy==1.23.5
opencv-python==4.8.0.74
tensorflow==2.12.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.9
stytch==9.3.0
stripe==7.8.0
